backend/requirements.txt
diffusers==0.21.4
transformers==4.31.0
huggingface-hub==0.16.4
accelerate==0.21.0
fastapi==0.95.2
python-multipart==0.0.6
pillow==9.5.0
uvicorn==0.22.0
numpy==1.24.3
scikit-learn==1.2.2
python-dotenv==1.0.0